Simple python scripts to record footage and routinely delete old footage
basic python script to record camera footage to a SD card directory and it creates a new file every 60 seconds. Once 75% capacity is reached, it will purge. Tweaking on the resolution and possibly finding another library for encoding could be beneficial, but I just wanted something to work at the moment.