Fixed socket_bytes_to_recv() bug
On x86/amd64 systems, using ssize_t with ioctl() returns a seemingly random number while on armf systems this works. Change the implementation of socket_bytes_to_recv() to use int instead of ssize_t, which is apparently the proper type to use but was not directly said in documentation. Updated all code that use and indirectly use this function.
* Also fixed if statement for sending a long test string in test_socket_srv.c that would cause the test program to fail when it should. a less than sign should have been a greater than sign.