Fixed bug that make socket_receive_data() return incorrect number bytes
On the RPi units, it appears the using ioctl(socket_fd,FIONREAD,&bytes) to get the number of bytes to be received by a socket results in a random and big number being returned. Had to change the socket_receive_data() to rely on read() to return the correct number of bytes instead of ioctl().

Added another test case to socket test services