Don't use hostnames
Just incase there's problems with hostnames, just send wlan0 IP instead since that won't rely on whatever the main unit is using to resolve hostnames. Discover packet still sends back the hostname, but it's only used for naming purposes in the GUI so it doesn't matter if it can't be resolved since the IP is sent back.